Update

update Update Update the results from drscore for continuous data only.

update(DRisk, ...)

Arguments

DRisk

results from drscore. Drisk object.

...

other arguments to drscore

Value

An object of the type as `DRisk`. Summary of the disclosure risk in terms of distinctness, similarities and undeniable parameters.

Examples

# ##################
# ##all continuous###################

library(svMisc)
library(dress)
library(sdcMicro)

CASC_sample <- CASCrefmicrodata[,c(2,3,4,6)]
CASC_protected <- addNoise(CASC_sample,noise = 100)$xm #Additive Noise protected

DRisk_NN <- drscore(
  Sample = CASC_sample, #Original Sample
  Protected = CASC_protected,
  delta = 0.05,
  kdistinct = 0.05, #k distinct threshold if integer then
  # probability threshold is k/SS (SS = sample size)
  ldeniable = 5, # l undeniable threshold if integer then
  # probability threshold is l/SS (SS = sample size)
  neighbourhood = 1,
  #Possible 'neighbourhood' types
  # 1 = Mahalanobis (Based on Mahalanobis Distance)
  # 2 = DSTAR   (Based on Density Based Distance)
  # 3 = StdEuclid (Based on Standardised (by std dev) Euclidean Distance)
  # 4 = RelEuclid (Relative Euclidean Distance sum_k ((Xk-Yk)/Xk)^2)
  neigh_type = 'prob',
  #Possible 'neigh_type' types
  #constant = fixed threshold on distance
  #prob = Nearest Neighbour Probability Neighbourhood used
  #estprob = = Nearest Neighbour Probability Neighbourhood used based on protected density
  numeric.vars = 1:4, #Which Variables are continuous?
  outlier.par = list(centre = median,
                     scale = var,
                     thresh = 0.01)
  #Parameters to adjust how MV outliers are determined.
  #Default is that lie 99% (based on Chi-Square n-1 dist) away from median after scale by variance.
)
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> 
#> ###################################################################### 
#> #                     Disclosure Risk Assessment                     # 
#> ###################################################################### 
#> Nearest Neighbour Neighbourhood with parameters:
#>         delta = 0.05, kdistinct = 0.05, ldeniable = 0.00462962962962963. 
#> 
#> Number of Observations in the Sample                        1080
#> Number of Observations in the Protected Sample              1080
#> Number of Continuous Variables                              4
#> Number of Key Categories                                    1
#> Number of Outliers in Sample                                38
#> Number of Distinct Points in Sample                         1080
#> Number of Distinct Outliers in Sample                       38
#> Number of Exact Matches in Sample                           0
#> Number of Interval Matches in Sample                        0
#> Number of Outlier Interval Matches in Sample                0
#> Number of Distint Outlier Interval Matches in Sample        0 
#>  
#> Delta Disclosure Risk of Sample                             0.0954
#> Delta Disclosure Risk of Sample Outliers                    0.7368
#> Proportion Distinct                                         1
#> Proportion Estimated                                        0.9509
#> Proportion Undeniable                                       0.0954 
#>  
#> Category Level Disclosure Risk: 
#>  
#>     N.Obs      DRisk  Out_DRisk Distinct Estimated Undeniable
#> All  1080 0.09537037 0.02592593        1 0.9509259 0.09537037
#> 

#Update neighbourhood to fixed threshold definition
DRisk_Fxd <- update(DRisk_NN,neigh_type = 'constant',
                    delta = 1)
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> 
#> ###################################################################### 
#> #                     Disclosure Risk Assessment                     # 
#> ###################################################################### 
#> Threshold Neighbourhood with parameters:
#>         delta = 1, kdistinct = 0.05, ldeniable = 0.00462962962962963. 
#> 
#> Number of Observations in the Sample                        1080
#> Number of Observations in the Protected Sample              1080
#> Number of Continuous Variables                              4
#> Number of Key Categories                                    1
#> Number of Outliers in Sample                                38
#> Number of Distinct Points in Sample                         642
#> Number of Distinct Outliers in Sample                       38
#> Number of Exact Matches in Sample                           0
#> Number of Interval Matches in Sample                        3
#> Number of Outlier Interval Matches in Sample                0
#> Number of Distint Outlier Interval Matches in Sample        0 
#>  
#> Delta Disclosure Risk of Sample                             0.1537
#> Delta Disclosure Risk of Sample Outliers                    0.3421
#> Proportion Distinct                                         0.5944
#> Proportion Estimated                                        0.9583
#> Proportion Undeniable                                       0.1574 
#>  
#> Category Level Disclosure Risk: 
#>  
#>     N.Obs     DRisk  Out_DRisk  Distinct Estimated Undeniable
#> All  1080 0.1537037 0.01203704 0.5944444 0.9583333  0.1574074
#>