wage4_protected

Wage and other data protected with some disclosure control method for original the Mid-Atlantic region wage data.

wage4_protected

Format

A data frame with 3000 observations and 4 variables:

age

Age of worker

education

A factor with levels 1. < HS Grad 2. HS Grad 3. Some College 4. College Grad and 5. Advanced Degree indicating education level

jobclass

A factor with levels 1. Industrial and 2. Information indicating type of job

wage

Workers raw wage

...

Source

Data was manually assembled by Steve Miller, of Inquidia Consulting (formerly Open BI). From the March 2011 Supplement to Current Population Survey data. 'https://www.re3data.org/repository/r3d100011860'

References

James, G., Witten, D., Hastie, T., and Tibshirani, R. (2013) An Introduction to Statistical Learning with applications in R, https://www.statlearning.com, Springer-Verlag, New York

See also

Examples

head(wage4_protected)
#>             age       education       jobclass      wage
#> 231655 47.58090 4. College Grad  1. Industrial  28.01377
#> 86582  34.51521 3. Some College 2. Information 193.61852
#> 161300 21.78810    1. < HS Grad  1. Industrial 116.36193
#> 155159 59.03079      2. HS Grad  1. Industrial  71.97888
#> 11443  27.26570 3. Some College 2. Information 171.46954
#> 376662 30.45184      2. HS Grad 2. Information  74.33412